+++ src/SPlot2d/SPlot2d_ViewModel.cxx
@@ -102,7 +102,7 @@ void SPlot2d_Viewer::renameAll( const Handle(SALOME_InteractiveObject)& IObj, co
   {
     const QVector<SUIT_ViewWindow*>& wnds = vm->getViews();
 
-    for ( uint i = 0; i < wnds.size(); i++ )
+    for ( uint i = 0; i < (uint)wnds.size(); i++ )
     {
       Plot2d_ViewWindow* pwnd = dynamic_cast<Plot2d_ViewWindow*>( wnds.at( i ) );
       rename( IObj, name, pwnd->getViewFrame() );
