+++ src/SalomeApp/SalomeApp_StudyPropertiesDlg.cxx
@@ -225,7 +225,7 @@ void SalomeApp_StudyPropertiesDlg::initData()
     std::vector<std::string> aUsers;
     std::vector<int>  aMins, aHours, aDays, aMonths, aYears;
     propAttr->GetModificationsList(aUsers, aMins, aHours, aDays, aMonths, aYears, false);
-    int aCnt = aUsers.size();
+    size_t aCnt = aUsers.size();
     for ( int i = 0; i < aCnt; i++ ) {
       QString date;
       date.sprintf("%2.2d/%2.2d/%2d %2.2d:%2.2d",
